eval/hyde/requirements.txt
openai == 1.55.3
scikit-learn == 1.5.2
python-dotenv == 1.0.1